Refactor to fix up warnings.

+++ include/bitcoin/database/databases/transaction_database.hpp
@@ -44,7 +44,7 @@ public:
 
     /// Construct the database.
     transaction_database(const path& map_filename, size_t table_minimum,
-        size_t buckets, size_t expansion, size_t cache_capacity);
+        uint32_t buckets, size_t expansion, size_t cache_capacity);
 
     /// Close the database (all threads must first be stopped).
     ~transaction_database();
